Move license-related hooks to hooks/license
Add also hook to check for availability of request update with the
current license.
