+++ webapp/src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './crud';
 export * from './routing';
 export * from './permissions';
 export * from './telemetry';
+export * from './license';
